Code structure refactor.
