+++ src/app/add-user/components/index.ts
@@ -1,0 +1,1 @@
+export { AddUserComponent } from './add-user/add-user.component';
