+++ src/app/add-user/index.ts
@@ -1,0 +1,1 @@
+export { AddUserModule } from './add-user.module';
