+++ src/app/menu/components/index.ts
@@ -1,0 +1,1 @@
+export { MenuComponent } from './menu/menu.component';
