+++ src/app/menu/index.ts
@@ -1,0 +1,1 @@
+export { MenuModule } from './menu.module';
