+++ src/app/user-list/components/index.ts
@@ -1,0 +1,1 @@
+export { UserListComponent } from './user-list/user-list.component';
