+++ src/app/user-list/index.ts
@@ -1,0 +1,1 @@
+export { UserListModule } from './user-list.module';
