+++ src/app/user-list/services/index.ts
@@ -1,0 +1,1 @@
+export { UserService } from './user-service';
