+++ src/app/user-list/user-list-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { UserListComponent } from './user-list.component';
+import { UserListComponent } from './components';
 
 const routes: Routes = [
     {
