+++ src/app/user-list/user-list.module.ts
@@ -7,7 +7,7 @@ import { MatTableModule } from '@angular/material/table';
 import { MatSortModule } from '@angular/material/sort';
 
 import { UserListRoutingModule } from './user-list-routing.module';
-import { UserListComponent } from './user-list.component';
+import { UserListComponent } from './components';
 
 @NgModule({
     declarations: [UserListComponent],
